Rename indexer_manager module to indexer
