+++ scripts/debug_lazy_loading.py
@@ -3,7 +3,7 @@ import time
 from pathlib import Path
 
 from pycontextify.config import Config
-from pycontextify.indexer_manager import IndexManager
+from pycontextify.indexer import IndexManager
 
 
 def debug_index_manager():
