+++ scripts/fast_startup_test.py
@@ -3,7 +3,7 @@ import time
 from pathlib import Path
 
 from pycontextify.config import Config
-from pycontextify.indexer_manager import IndexManager
+from pycontextify.indexer import IndexManager
 
 
 def test_fast_startup_config():
