+++ tests/integration/test_persistence.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 from pycontextify.config import Config
-from pycontextify.indexer_manager import IndexManager
+from pycontextify.indexer import IndexManager
 
 
 class TestPersistenceFeatures:
